refactor: add coffee sizes.
The solution suggested by the text book seems to violate the decorator
pattern itself (the pattern to be learned with this example). Imagine
an additional cup size is added to the menu, all of the existing decorator
implementations must be re-touched to account for the new size. This
violates the Open-Closed principle.